pyproject.toml
[tool.poetry]
name = "data-spectrum"
version = "0.1.0"
description = "Universal Metadata Extraction Tool for Databases, Data Lakes, and File Systems"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.25"
cryptography = "^42.0.0"
fsspec = "^2023.12.0"
s3fs = "^2023.12.0"
adlfs = "^2023.12.0"
gcsfs = "^2023.12.0"
pyarrow = "^15.0.0"
deltalake = "^0.15.0"
pyiceberg = "^0.5.0"
databricks-sql-connector = "^3.0.0"
snowflake-sqlalchemy = "^1.5.0"
pyyaml = "^6.0.1"
duckdb = "^0.9.2"
databricks-sdk = "^0.80.0"
duckdb-engine = "^0.17.0"
typer = "^0.21.1"
rich = "<14.0.0"
prompt-toolkit = "^3.0.52"

[tool.poetry.scripts]
data-spectrum = "src.cli.main:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
httpx = "^0.26.0"
testcontainers = "^4.14.0"
psycopg2-binary = "^2.9.11"
minio = "^7.2.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
